Ensure UNIX implementations of PltGetMillis() are not limited to 32 bits

+++ src/Platform.c
@@ -401,13 +401,13 @@ uint64_t PltGetMillis(void) {
 
     clock_gettime(CLOCK_MONOTONIC, &tv);
 
-    return (tv.tv_sec * 1000) + (tv.tv_nsec / 1000000);
+    return ((uint64_t)tv.tv_sec * 1000) + (tv.tv_nsec / 1000000);
 #else
     struct timeval tv;
 
     gettimeofday(&tv, NULL);
 
-    return (tv.tv_sec * 1000) + (tv.tv_usec / 1000);
+    return ((uint64_t)tv.tv_sec * 1000) + (tv.tv_usec / 1000);
 #endif
 }
